Fixed a bug in function call inlining.
The problem was that the parameters of a function call were clashing
with the caller context --- a classic bug in function inlining. To fix
this we take the called code and make it distinct from the
free-variables of the calling context.

Example of the expected behavior follows:

Function: f(x,y) { decl z = x + y; return z; }
Context: existing variables x, y
Code: decl g = f(1, 2); A[x + y];

Expected result:
  decl x1 = 1;
  decl y1 = 2;
  decl g = x1 + y1;
  decl z = x1 + y1;
  A[x + y];